% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3_january.R
\name{create.dXdT}
\alias{create.dXdT}
\title{create.dXdt}
\usage{
create.dXdT(
  case.data,
  time.var,
  timepoint.id = NULL,
  case.id = NULL,
  features,
  scheme
)
}
\arguments{
\item{case.data, time.var, timepoint.id, case.id, features, scheme}{Data frame with all observations,
variable name or index in case.data for time observation,
variable name or index in case.data for time point identifiers,
variable name or index in case.data for cases,
a data frame or matrix with all the features,
scheme for calculating dT.}
}
\value{
Return a DxDt object.
}
\description{
create.dXdt
}
