% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3_january.R
\name{create.dXdTdist}
\alias{create.dXdTdist}
\title{create.dXdtdist}
\usage{
create.dXdTdist(dm)
}
\arguments{
\item{dm}{A distance matrix of features between each observation.}
}
\value{
Return a DxDt object using principal component analysis.
}
\description{
create.dXdtdist
}
