% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md3_january.R
\name{process_data}
\alias{process_data}
\title{process_data}
\usage{
process_data(data, time_scheme)
}
\arguments{
\item{data, time_scheme}{Data to be processed, two schemes, baseline and interpoint}
}
\value{
Return processed data based on scheme of calculation
}
\description{
process_data
}
